Add pod uid to the pod registry key
The controller clears the data related to a pod when it is
deleted/restarted from the pod registry ets table. But the key
that is represented by the pod name is still in the ets table.

This may lead to the situation when start of new pod will be handled
incorrectly, as for example if the new pod name is the same that
previous as in statefulsets.

This commit adds pod uid to the key to make keys uniq regardless
pod names.